requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
Flask==3.0.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.53.0
pluggy==1.6.0
psutil==7.0.0
psycopg==3.1.18
pyee==13.0.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
pytest-xprocess==0.23.0
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
